dm cache policy smq: add a tunable for populating cache aggressively
In some cases where the working set is consistently smaller than the
cache size, users may prefer to always promote missed data blocks to
quickly achieve a hit state during workload execution. The default smq
policy doesn't support this requirement as it needs multiple misses on a
single block to determine what to promote.

To address this issue, a new tunable parameter "fast_populate" is
introduced. By setting the parameter to 1, the smq policy considers
every missed block as promotable while the cache is not full, thereby
populating the cache more aggressively. Once the cache is fully
populated, the smq policy maintains the promotion threshold in the
usual way.